add validations

+++ assets/helpers/toDateFormat.ts
@@ -1,7 +1,7 @@
 export function toDateFormat(date: string): string {
   return new Date(date + `Z`).toLocaleString('ru-RU', {
     day: 'numeric',
-    month: 'long',
+    month: 'numeric',
     year: 'numeric',
   });
 }
